+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [1.2.1]
 ### Added
 - More context to error notifications.
 - Separate Gitlab url and token table.